+++ R/mod_input_filter.R
@@ -14,11 +14,13 @@ mod_input_filter_ui <- function(id) {
     shiny::div(
       id = id,
       shiny::uiOutput(ns("filter_var_ui")),
+      shiny::br(),
       shiny::uiOutput(ns("filter_val_ui")),
+      shiny::br(),
       tags$style(shiny::HTML(paste0("#", id, "-filter_var_ui{display:inline-block}"))),
       tags$style(shiny::HTML(paste0("#", id, "-filter_val_ui{display:inline-block; vertical-align: top;}"))),
-      tags$style(shiny::HTML(paste0("#", id, "-filter_min_val_sel{display:inline-block}"))), # TODO: make these conditional html output
-      tags$style(shiny::HTML(paste0("#", id, "-filter_max_val_sel{display:inline-block}")))
+      tags$style(shiny::HTML(paste0("#", id, "-filter_min_val_sel{display:inline-block; vertical-align: top;}"))), # TODO: make these conditional html output
+      tags$style(shiny::HTML(paste0("#", id, "-filter_max_val_sel{display:inline-block; vertical-align: top;}")))
     )
   )
 }
@@ -48,6 +50,12 @@ mod_input_filter_server <- function(id, data, filter_opts, all_filters, n_filter
       class(data()[[input$filter_var_sel]]) == "Date"
     })
 
+    # This can be transformed is_numeric, if we allow other numerical variables to the filtering in some point
+    is_age <- shiny::reactive({
+      shiny::req(input$filter_var_sel)
+      input$filter_var_sel == "age"
+    })
+
     # filter value
     output$filter_val_ui <- shiny::renderUI({
       shiny::req(input$filter_var_sel != "None")
@@ -69,6 +77,28 @@ mod_input_filter_server <- function(id, data, filter_opts, all_filters, n_filter
             max = max(data()[[input$filter_var_sel]]),
             value = max(data()[[input$filter_var_sel]])
           ),
+        )
+      } else if(is_age()) {
+
+        value_ui <- shiny::tagList(
+          shiny::div(
+            shiny::numericInput(
+              inputId = ns("filter_min_val_sel"),
+              label = "Minimum Age",
+              value = min(data()[[input$filter_var_sel]], na.rm = T),
+              min = min(data()[[input$filter_var_sel]], na.rm = T),
+              max = max(data()[[input$filter_var_sel]], na.rm = T)
+              ),
+            style = "display: inline-block; width: 45%; vertical-align: top;"),
+          shiny::div(
+            shiny::numericInput(
+              inputId = ns("filter_max_val_sel"),
+              label = "Maximum Age",
+              min = min(data()[[input$filter_var_sel]], na.rm = T),
+              max = max(data()[[input$filter_var_sel]], na.rm = T),
+              value = max(data()[[input$filter_var_sel]], na.rm = T)
+            ),
+            style = "display: inline-block; width: 45%; vertical-align: top;")
         )
       } else {
         if (class(data()[[input$filter_var_sel]]) == "factor") {
@@ -95,12 +125,48 @@ mod_input_filter_server <- function(id, data, filter_opts, all_filters, n_filter
     # return values to filter for
     filter_val <- shiny::reactive({
       shiny::req(is_date)
+      shiny::req(is_age)
       if (is_date()) {
+        values <- c(input$filter_min_val_sel, input$filter_max_val_sel)
+      } else if(is_age()){
         values <- c(input$filter_min_val_sel, input$filter_max_val_sel)
       } else {
         values <- input$filter_val_sel
       }
       values
+    })
+
+    # Logic for filtering for age
+    shiny::observeEvent(input$filter_min_val_sel, {
+      # Ensure the input is not empty or NULL
+      if (!is.null(input$filter_min_val_sel) && !is.null(input$filter_max_val_sel)) {
+        min_val <- input$filter_min_val_sel
+        max_val <- input$filter_max_val_sel
+
+        # If the minimum value exceeds the maximum, reset it to the maximum
+        if (!is.na(min_val) && min_val > max_val) {
+          updateNumericInput(session, "filter_min_val_sel", value = max_val)
+        } else {
+          # Otherwise, just update the max limit for the minimum input
+          updateNumericInput(session, "filter_max_val_sel", min = min_val)
+        }
+      }
+    })
+
+    shiny::observeEvent(input$filter_max_val_sel, {
+      # Ensure the input is not empty or NULL
+      if (!is.null(input$filter_min_val_sel) && !is.null(input$filter_max_val_sel)) {
+        min_val <- input$filter_min_val_sel
+        max_val <- input$filter_max_val_sel
+
+        # If the maximum value is less than the minimum, reset it to the minimum
+        if (!is.na(max_val) && max_val < min_val) {
+          updateNumericInput(session, "filter_max_val_sel", value = min_val)
+        } else {
+          # Otherwise, just update the min limit for the maximum input
+          updateNumericInput(session, "filter_min_val_sel", max = max_val)
+        }
+      }
     })
 
     # update filter_var choices depending on all_filters (n_filters is required
@@ -125,7 +191,6 @@ mod_input_filter_server <- function(id, data, filter_opts, all_filters, n_filter
         )
       }
     )
-
 
     # return
     list(